Use floats for filtering

+++ Drivers/generalIO/adc_global.h
@@ -20,9 +20,9 @@ typedef struct{
   volatile int16_t    prev_avg;
   volatile int16_t    prev_raw;
   #endif
-  volatile int16_t    last_avg;               // Filtered (EMA calculation)
-  volatile int16_t    last_raw;               // Unfiltered, for quick Iron detection
-  volatile uint32_t   EMA_of_Input;           // Stored filter data (acumulator for EMA)
+  volatile uint16_t   last_avg;               // Filtered (EMA calculation)
+  volatile uint16_t   last_raw;               // Unfiltered, for quick Iron detection
+  volatile float      EMA_of_Input;           // Stored filter data (acumulator for EMA)
   volatile uint16_t   *adc_buffer;            // Ptr to ADC buffer data
 
 } ADCDataTypeDef_t;
